requirements.txt
discord
youtube_dl
asyncio
python-decouple
PyNaCl